Added some revisions to the code

+++ priprema za pismeni i usmeni ispit/125. zadatak/a)/a).c
@@ -5,7 +5,7 @@ float rekurzivna(int x)
     if (x == 1)
         return 1;
     else
-        return rekurzivna(x - 1) + 1 / (float)(x);
+        return rekurzivna(x - 1) + 1. / x;
 }
 
 int main(void)
